Refactored app out server.js; added basic backend test
