+++ backend/package.json
@@ -1,21 +1,22 @@
 {
   "dependencies": {
-    "jest": "^27.3.1",
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.1",
+    "cross-env": "^7.0.3",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "jest": "^27.3.1",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.9.1",
+    "supertest": "^6.3.3",
     "validator": "^13.9.0"
   },
   "scripts": {
-    "test": "jest",
+    "test": "cross-env NODE_ENV=test jest --testTimeout=5000",
     "start": "node server.js",
     "server": "nodemon server.js"
   },
   "devDependencies": {
-    "nodemon": "^2.0.20",
     "eslint": "^8.34.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-standard": "^17.0.0",
@@ -23,6 +24,7 @@
     "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-n": "^15.6.1",
     "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "^7.32.2"
+    "eslint-plugin-react": "^7.32.2",
+    "nodemon": "^2.0.20"
   }
 }
